Move classes and packages
This moves classes and packages to create a more logical hierarchy.
