+++ relution-publisher/src/main/java/org/jenkinsci/plugins/relution_publisher/builder/ArtifactFileUploader.java
@@ -25,13 +25,13 @@ import org.apache.commons.lang.StringUtils;
 import org.apache.tools.ant.types.FileSet;
 import org.jenkinsci.plugins.relution_publisher.configuration.global.Store;
 import org.jenkinsci.plugins.relution_publisher.configuration.jobs.Publication;
-import org.jenkinsci.plugins.relution_publisher.constants.ApiObject;
-import org.jenkinsci.plugins.relution_publisher.constants.App;
-import org.jenkinsci.plugins.relution_publisher.constants.ArchiveMode;
-import org.jenkinsci.plugins.relution_publisher.constants.Language;
-import org.jenkinsci.plugins.relution_publisher.constants.Version;
 import org.jenkinsci.plugins.relution_publisher.logging.Log;
+import org.jenkinsci.plugins.relution_publisher.model.ArchiveMode;
 import org.jenkinsci.plugins.relution_publisher.model.ServerVersion;
+import org.jenkinsci.plugins.relution_publisher.model.entities.ApiObject;
+import org.jenkinsci.plugins.relution_publisher.model.entities.App;
+import org.jenkinsci.plugins.relution_publisher.model.entities.Language;
+import org.jenkinsci.plugins.relution_publisher.model.entities.Version;
 import org.jenkinsci.plugins.relution_publisher.net.RequestFactory;
 import org.jenkinsci.plugins.relution_publisher.net.SessionManager;
 import org.jenkinsci.plugins.relution_publisher.net.requests.ApiRequest;
