+++ relution-publisher/src/main/java/org/jenkinsci/plugins/relution_publisher/configuration/global/Store.java
@@ -25,10 +25,10 @@ import org.apache.commons.validator.routines.UrlValidator;
 import org.apache.http.HttpException;
 import org.apache.http.HttpStatus;
 import org.apache.http.ParseException;
-import org.jenkinsci.plugins.relution_publisher.constants.ArchiveMode;
-import org.jenkinsci.plugins.relution_publisher.constants.ReleaseStatus;
-import org.jenkinsci.plugins.relution_publisher.constants.UploadMode;
+import org.jenkinsci.plugins.relution_publisher.model.ArchiveMode;
+import org.jenkinsci.plugins.relution_publisher.model.ReleaseStatus;
 import org.jenkinsci.plugins.relution_publisher.model.ServerVersion;
+import org.jenkinsci.plugins.relution_publisher.model.UploadMode;
 import org.jenkinsci.plugins.relution_publisher.net.RequestFactory;
 import org.jenkinsci.plugins.relution_publisher.net.SessionManager;
 import org.jenkinsci.plugins.relution_publisher.net.requests.BaseRequest;
