+++ relution-publisher/src/main/java/org/jenkinsci/plugins/relution_publisher/configuration/jobs/Publication.java
@@ -19,9 +19,9 @@ package org.jenkinsci.plugins.relution_publisher.configuration.jobs;
 import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.plugins.relution_publisher.configuration.global.Store;
 import org.jenkinsci.plugins.relution_publisher.configuration.global.StoreConfiguration;
-import org.jenkinsci.plugins.relution_publisher.constants.ArchiveMode;
-import org.jenkinsci.plugins.relution_publisher.constants.ReleaseStatus;
-import org.jenkinsci.plugins.relution_publisher.constants.UploadMode;
+import org.jenkinsci.plugins.relution_publisher.model.ArchiveMode;
+import org.jenkinsci.plugins.relution_publisher.model.ReleaseStatus;
+import org.jenkinsci.plugins.relution_publisher.model.UploadMode;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.QueryParameter;
